change annotations

+++ deploy/run_teleoperation_mujoco.py
@@ -1,5 +1,5 @@
 from deploy.config import Config
-from deploy.controllers.controller import Runner_handle_mujoco_vision #, Runner_offline_mujoco
+from deploy.controllers.controller import Runner_handle_mujoco_vision  #, Runner_offline_mujoco
 
 import time
 import cv2
@@ -14,25 +14,29 @@ from deploy.teleop.image_server.image_client import ImageClient
 
 import numpy as np
 import torch
+
 torch.set_printoptions(precision=3)
 np.set_printoptions(precision=3)
 
 tv_img_shape = (480, 1280, 3)
 tv_img_dtype = np.uint8
-tv_img_shm = shared_memory.SharedMemory(create = True, size = np.prod(tv_img_shape) * np.uint8().itemsize)
-tv_img_array = np.ndarray(tv_img_shape, dtype = tv_img_dtype, buffer = tv_img_shm.buf)
+tv_img_shm = shared_memory.SharedMemory(create=True, size=np.prod(tv_img_shape) * np.uint8().itemsize)
+tv_img_array = np.ndarray(tv_img_shape, dtype=tv_img_dtype, buffer=tv_img_shm.buf)
 
 # television: obtain hand pose data from the XR device and transmit the robot's head camera image to the XR device.
 tv_wrapper = TeleVisionWrapper(True, tv_img_shape, tv_img_shm.name)
 
 arm_ik = G1_29_ArmIK()
+
+
 def get_output_dir():
     timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
     output_dir = os.path.join("record_mujoco_images", timestamp)
     os.makedirs(output_dir, exist_ok=True)
     return output_dir
 
-def save_images(shm_name, shape, dtype, interval=1.0/25.):
+
+def save_images(shm_name, shape, dtype, interval=1.0 / 25.):
     shm = shared_memory.SharedMemory(name=shm_name)
     img = np.ndarray(shape, dtype=dtype, buffer=shm.buf)
 
@@ -51,6 +55,7 @@ def save_images(shm_name, shape, dtype, interval=1.0/25.):
     finally:
         shm.close()
 
+
 def deploy_handle_mujoco(args):
     import mujoco.viewer
     import mujoco
@@ -58,12 +63,12 @@ def deploy_handle_mujoco(args):
     print(config_path)
     config = Config(config_path)
 
-
     # Initialize DDS communication
     runner = Runner_handle_mujoco_vision(config, args=args)
     current_mode = "SQUAT"
     print('Squat mode!')
     print('Press Left_A to start the locomotion mode!')
+
     def tv_arms():
         head_rmat, left_wrist, right_wrist, left_hand, right_hand = tv_wrapper.get_data()
         current_lr_arm_q = runner.qj.copy()[15:29]
@@ -73,6 +78,7 @@ def deploy_handle_mujoco(args):
         sol_q = np.clip(sol_q, runner.target_dof_pos[runner.config.action_hl_idx] - 0.01,
                         runner.target_dof_pos[runner.config.action_hl_idx] + 0.01)
         runner.target_dof_pos[runner.config.action_hl_idx] = sol_q
+
     if args.save_image:
         p_record_video = Process(target=save_images, args=(tv_img_shm.name, tv_img_shape, tv_img_dtype))
         p_record_video.start()
@@ -107,11 +113,15 @@ def deploy_handle_mujoco(args):
     tv_img_shm.unlink()
     tv_img_shm.close()
 
+
 if __name__ == "__main__":
     import argparse
 
     parser = argparse.ArgumentParser()
-    parser.add_argument("--config", type=str, help="config file name in the configs folder", default="run_loco_squat_grasp.yaml")
+    parser.add_argument("--config",
+                        type=str,
+                        help="config file name in the configs folder",
+                        default="run_loco_squat_grasp.yaml")
     parser.add_argument("--save_data", action="store_true", help="whether saving the mujoco data")
     parser.add_argument("--save_data_dir", type=str, help="where to save the data", default="./save_mujoco_data")
     parser.add_argument("--save_image", action="store_true", help="whether saving the mujoco image")
