+++ deploy/run_teleoperation_real.py
@@ -1,7 +1,6 @@
-
 # import onnxruntime as ort
 from deploy.config import Config
-from deploy.controllers.controller import Runner_online_real_dexhand, Runner_handle_mujoco #, Runner_offline_mujoco
+from deploy.controllers.controller import Runner_online_real_dexhand, Runner_handle_mujoco  #, Runner_offline_mujoco
 
 import time
 import cv2
@@ -16,29 +15,34 @@ from deploy.teleop.image_server.image_client import ImageClient
 
 import numpy as np
 import torch
+
 torch.set_printoptions(precision=3)
 np.set_printoptions(precision=3)
 
 tv_img_shape = (480, 1280, 3)
 tv_img_dtype = np.uint8
-tv_img_shm = shared_memory.SharedMemory(create = True, size = np.prod(tv_img_shape) * np.uint8().itemsize)
-tv_img_array = np.ndarray(tv_img_shape, dtype = tv_img_dtype, buffer = tv_img_shm.buf)
-img_client = ImageClient(tv_img_shape = tv_img_shape, tv_img_shm_name = tv_img_shm.name) # , server_address='10.100.6.192', port=8012
+tv_img_shm = shared_memory.SharedMemory(create=True, size=np.prod(tv_img_shape) * np.uint8().itemsize)
+tv_img_array = np.ndarray(tv_img_shape, dtype=tv_img_dtype, buffer=tv_img_shm.buf)
+img_client = ImageClient(tv_img_shape=tv_img_shape,
+                         tv_img_shm_name=tv_img_shm.name)  # , server_address='10.100.6.192', port=8012
 
-image_receive_thread = threading.Thread(target = img_client.receive_process, daemon = True)
+image_receive_thread = threading.Thread(target=img_client.receive_process, daemon=True)
 image_receive_thread.start()
 
 # television: obtain hand pose data from the XR device and transmit the robot's head camera image to the XR device.
 tv_wrapper = TeleVisionWrapper(True, tv_img_shape, tv_img_shm.name)
 
 arm_ik = G1_29_ArmIK()
+
+
 def get_output_dir():
     timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
     output_dir = os.path.join("record_real_images", timestamp)
     os.makedirs(output_dir, exist_ok=True)
     return output_dir
 
-def save_images(shm_name, shape, dtype, interval=1.0/25.):
+
+def save_images(shm_name, shape, dtype, interval=1.0 / 25.):
     shm = shared_memory.SharedMemory(name=shm_name)
     img = np.ndarray(shape, dtype=dtype, buffer=shm.buf)
 
@@ -57,6 +61,7 @@ def save_images(shm_name, shape, dtype, interval=1.0/25.):
     finally:
         shm.close()
 
+
 def deploy_real(args):
     # Load config
     config_path = f"deploy/configs/{args.config}"
@@ -72,6 +77,7 @@ def deploy_real(args):
     current_mode = "SQUAT"
     print('Squat mode!')
     print('Press Left_A to start the locomotion mode!')
+
     def tv_arms():
         head_rmat, left_wrist, right_wrist, left_hand, right_hand = tv_wrapper.get_data()
         current_lr_arm_q = runner.qj.copy()[15:29]
@@ -92,7 +98,7 @@ def deploy_real(args):
                 current_mode = "SQUAT"
                 print('Squat mode!')
                 print('Press Left_A to start the locomotion mode!')
-    
+
         elif current_mode == "SQUAT":
             tv_arms()
             runner.run_squat_hand(debug=args.debug, manual=True)
@@ -112,12 +118,16 @@ def deploy_real(args):
     print("Finally, exiting program...")
     exit(0)
 
+
 if __name__ == "__main__":
     import argparse
 
     parser = argparse.ArgumentParser()
     parser.add_argument("--net", type=str, help="network interface")
-    parser.add_argument("--config", type=str, help="config file name in the configs folder", default="run_loco_squat_grasp.yaml")
+    parser.add_argument("--config",
+                        type=str,
+                        help="config file name in the configs folder",
+                        default="run_loco_squat_grasp.yaml")
     parser.add_argument("--save_data", action="store_true", help="whether saving the real data")
     parser.add_argument("--save_data_dir", type=str, help="where to save the data", default="./save_real_data")
     parser.add_argument("--save_image", action="store_true", help="whether saving the real image")
